download script run
